Implemented Event saving

+++ EventHub/ViewModels/EventFormViewModel.cs
@@ -1,8 +1,5 @@
 ﻿using EventHub.Models;
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Web;
 
 namespace EventHub.ViewModels
 {
@@ -11,7 +8,7 @@ namespace EventHub.ViewModels
         public string Venue { get; set; }
         public string Date { get; set; }
         public string Time { get; set; }
-        public int Topic { get; set; }
+        public byte Topic { get; set; }
 
         public IEnumerable<Topic> Topics { get; set; }
     }
